melko valmis

+++ luokat.php
@@ -14,8 +14,29 @@ else {
 	
 		if (isset($_POST['poista'])) {
 			$luokka = $_POST['poista'];
-			$poisto = $yhteys->prepare("DELETE FROM luokka WHERE luokkaid=?");
-			$poisto->execute(array($luokka));
+			
+			$tarkista = $yhteys->prepare("SELECT nimi FROM luokka WHERE ylaluokka=?");
+			$tarkista->execute(array($luokka));
+			$onko_ylaluokkana = $tarkista->fetchObject();
+			
+			$tarkista = $yhteys->prepare("SELECT nimi FROM askare WHERE luokka=?");
+			$tarkista->execute(array($luokka));
+			$onko_luokkana = $tarkista->fetchObject();
+			
+			$tarkista = $yhteys->prepare("SELECT nimi FROM luokka WHERE luokkaid=?");
+			$tarkista->execute(array($luokka));
+			$lnimi = $tarkista->fetchObject()->nimi;
+			
+			if ($onko_ylaluokkana == TRUE) {
+				$virheteksti = 1;
+			}
+			else if ($onko_luokkana == TRUE) {
+				$virheteksti = 2;
+			}
+			else {
+				$poisto = $yhteys->prepare("DELETE FROM luokka WHERE luokkaid=?");
+				$poisto->execute(array($luokka));
+			}
 		} else if (isset($_POST['muokkaa'])) {
 			$_SESSION['l_id'] = $_POST['muokkaa'];
 			header('Location: uusi_luokka.php');
@@ -34,13 +55,13 @@ else {
 </head>
 
 <div class="headnav">
-  <a href="etusivu.php">Etusivu</a> * 
+  <a href="etusivu.php">Askareet</a> * 
   <a href="uusi_askare.php">Uusi askare</a> *  
   <a href="uusi_luokka.php">Uusi luokka</a> * 
   <a href="uloskirjautuminen.php">Kirjaudu ulos</a>
 </div>
 
-<div class="pienikehys">
+<div class="keskikehys">
 	<div class="ruutu">
 
 	<h2>Luokat</h2>	
@@ -49,7 +70,7 @@ else {
 	<table align="center">
 	<col width="150px"/>
 	<col width="150px"/>
-	<col width="60px"/>
+	<col width="55px"/>
 		<tr align="center">
 		<th class="noborder">Luokka</th>
 		<th class="noborder">Yläluokka</th>
@@ -82,7 +103,14 @@ else {
 	</table>
 	</p>
 	<p></p>
-
+		<?php
+		if ($virheteksti == 1) {
+			echo "Et voi poistaa luokkaa $lnimi, sillä se on käytössä yläluokkana.";
+		}
+		if ($virheteksti == 2) {
+			echo "Et voi poistaa luokkaa $lnimi, sillä se on käytössä yhden tai useamman askareen luokkana.";
+		}
+		?>
 </div>
 </div>
 <?php	} catch (PDOException $e) {
